routing problem solved
