+++ web/src/main/resources/public/assets/i18n/uk.json
@@ -3,7 +3,7 @@
   "apartments": "квартири",
   "HELLO": "privet",
   "ourYard": "НАШ ДВІР",
-  "providers": "провайдери",
+  "providers": "надавачі послуг",
   "company": "компанія",
   "type": "тип",
   "price": "ціна",
@@ -292,5 +292,6 @@
   "add_folder": "Створити папку",
   "folder_name": "Назва папки",
   "select_all": "Виділити все",
-  "rename": "Переіменувати"
+  "rename": "Переіменувати",
+  "upload_logo": "завантажити"
 }
