+++ web/src/main/resources/public/src/app/manager/components/provider/provider-table.html
@@ -1,10 +1,2 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>$Title$</title>
-</head>
-<body>
-$END$
-</body>
-</html>
+<myosbb-provider>
+</myosbb-provider>
