+++ web/src/main/resources/public/src/app/user/provider/provider.component.ts
@@ -1,10 +1,9 @@
 /**
  * Created by Anastasiia Fedorak  8/2/16.
  */
-import {Component, ViewChild, from
-"@angular/core";
+import {Component, ViewChild} from "@angular/core";
 import {CapitalizeFirstLetterPipe} from "../../../shared/pipes/capitalize-first-letter";
-import {TranslatePipe, TranslateService} from "ng2-translate/ng2-translate";
+import {TranslatePipe} from "ng2-translate/ng2-translate";
 import {DROPDOWN_DIRECTIVES} from "ng2-bs-dropdown/dropdown";
 import {ProviderService} from "./service/provider-service";
 import {PageCreator} from "../../../shared/services/page.creator.interface";
@@ -12,7 +11,7 @@ import {MODAL_DIRECTIVES, BS_VIEW_PROVIDERS, BUTTON_DIRECTIVES} from 'ng2-bootst
 import {ModalDirective} from "ng2-bootstrap/ng2-bootstrap";
 import {CORE_DIRECTIVES, NgClass} from "@angular/common";
 import {SELECT_DIRECTIVES} from "ng2-select/ng2-select";
-import {ROUTER_DIRECTIVES, Router} from "@angular/router";
+import {ROUTER_DIRECTIVES} from "@angular/router";
 import {ProviderTypeComponent} from "./provider_type/provider-type.component";
 import {Provider} from "../../../shared/models/provider.interface";
 import {Mail} from "../../../shared/models/mail.interface";
@@ -29,7 +28,7 @@ import FileLocationPath = require("../../../shared/services/file.location.path")
 import {AttachmentComponent} from "../../attachment/attachment.component";
 const attachmentUploadUrl = ApiService.serverUrl + '/restful/attachment/';
 
-declare var saveAs: any;
+declare let saveAs: any;
 
 
 @Component({
@@ -107,37 +106,21 @@ export class ProviderComponent {
     }
 
     ngOnInit(): any {
-        console.log("init provider cmp");
-        console.log("periodicity items:", PeriodicityItems);
         for (let i = 0; i < PeriodicityItems.length; i++) {
             this.periodicities.push(PeriodicityItems[i]);
         }
-
         this.getPeriodicitiesTranslation();
-        console.log('readable periodicities: ', this.periodicities);
         this.getProvidersByPageNumAndState(this.pageNumber);
     }
 
     setType(event) {
-        console.log("set type" + event);
         this.selected.type = event;
         this.newProvider.type = event;
-        console.log("selected.type=" + this.selected.type.providerTypeName + JSON.stringify(this.selected));
     }
 
     public onSelectPeriodicity(value: SelectItem): void {
-        console.log("value: ", value);
         this.selected.periodicity = this.backToConst(value);
         this.newProvider.periodicity = this.backToConst(value);
-        console.log("set periodicity: ", this.selected.periodicity);
-    }
-
-    public onRemove(value: SelectItem): void {
-        console.log('Removed value is: ', value);
-    }
-
-    public onType(value: SelectItem): void {
-        console.log('New search input: ', value);
     }
 
     public onRefresh(value: SelectItem): void {
@@ -148,12 +131,10 @@ export class ProviderComponent {
     openEditModal(provider: Provider) {
         this.selected = provider;
         this.upload = false;
-        console.log('selected provider: ' + this.selected);
         this.editModal.show();
     }
 
     closeEditModal() {
-        console.log('closing edt modal');
         this.editModal.hide();
         setTimeout(() => this.active = true, 0);
     }
@@ -162,15 +143,13 @@ export class ProviderComponent {
         this.active = false;
         if (this.shouldRun) {
             this._providerService.editProvider(this.selected).subscribe(() => {
-                    console.log("refreshing...");
                     this.shouldRun = false;
                     this.refresh();
                 },
                 (err) => {
-                    console.log(err)
+                    console.error(err)
                 }
             );
-            console.log("save provider: ", this.selected);
             this.editModal.hide();
             setTimeout(() => this.active = true, 0);
         }
@@ -180,7 +159,7 @@ export class ProviderComponent {
     openDelModal(provider: Provider) {
         this.selected = provider;
         this.upload = false;
-        console.log('selected provider: ' + this.selected);
+
         this.delModal.show();
     }
 
@@ -188,33 +167,28 @@ export class ProviderComponent {
         this.active = false;
         if (this.shouldRun) {
             this._providerService.makeProviderInactive(this.selected).subscribe(() => {
-                    console.log("refreshing...");
                     this.shouldRun = false;
                     this.refresh();
                 },
                 (err) => {
-                    console.log(err)
+                    console.error(err)
                 }
             );
-            console.log("making provider inactive: ", this.selected);
             this.delModal.hide();
         }
     }
 
     getPeriodicitiesTranslation() {
-        console.log("got lang", HeaderComponent.translateService.currentLang);
         for (let i = 0; i < this.periodicities.length; i++) {
             HeaderComponent.translateService.get(this.periodicities[i].text)
                 .subscribe((data: string) => {
                     this.periodicities[i].text = data;
-                    console.log("periodicity =", this.periodicities[i]);
                 })
         }
-        console.log("periodicities: ", this.periodicities);
     }
 
     backToConst(item: SelectItem): string {
-        var items: SelectItem[] =
+        let items: SelectItem[] =
             [{id: 1, text: 'ONE_TIME'},
                 {id: 2, text: 'PERMANENT_DAYLY'},
                 {id: 3, text: 'PERMANENT_WEEKLY'},
@@ -222,7 +196,6 @@ export class ProviderComponent {
                 {id: 5, text: 'PERMANENT_YEARLY'}];
         for (let i = 0; i < items.length; i++) {
             if (item.id === items[i].id) {
-                console.log("const: ", items[i].text);
                 return items[i].text;
             }
         }
@@ -251,9 +224,7 @@ export class ProviderComponent {
     }
 
     sortBy(name: string) {
-        console.log('sorted by ', name);
         this.order = !this.order;
-        console.log('order by asc', this.order);
         this.emptyArray();
         this._providerService.getSortedActiveProviders(this.pageNumber, name, this.order, this.onlyActive)
             .subscribe((data) => {
@@ -269,61 +240,46 @@ export class ProviderComponent {
     }
 
     onSearch(search: string) {
-        console.log("inside search: search param" + search);
         this._providerService.findProviderByNameOrDescription(search)
             .subscribe((providers) => {
-                console.log("data: " + providers);
                 this.providers = providers;
             });
     }
 
     onSendMessage() {
-        console.log("sending...");
         this.mail = {to: "aska.fed@gmail.com", subject: "TEST", text: "Success!"};
         this._mailService.sendMail(this.mail);
     }
 
     onCreateProviderSubmit() {
         this.active = false;
-        console.log("creating ", this.newProvider);
         if (this.shouldRun) {
             this._providerService.addProvider(this.newProvider)
                 .subscribe(() => {
-                        console.log("refreshing...");
                         this.shouldRun = false;
                         this.active = true;
                         this.refresh();
                         this.emptyFields();
                     },
                     (err) => {
-                        console.log(err)
+                        console.error(err)
                     }
                 );
         }
-
-        console.log("add provider", this.newProvider);
-
-
         let mail: Mail = {
             to: this.newProvider.email,
             subject: 'PRIVET',
             text: 'Welcome on the board'
         };
-        console.log("sending mail", mail);
         if (this.newProvider.email !== null) {
             this._mailService.sendMail(mail)
         }
-
-        console.log("DOUBLE FUCK")
-
         this.createModal.hide();
         setTimeout(() => this.active = true, 0);
     }
 
     closeCreateModal() {
-        console.log('closing create modal');
         this.createModal.hide();
-
     }
 
     openCreateModal() {
@@ -332,7 +288,6 @@ export class ProviderComponent {
     }
 
     getProvidersByPageNumAndState(pageNumber: number) {
-        console.log("getProvidersByPageNum " + pageNumber + "; only active=" + this.onlyActive);
         this.pageNumber = +pageNumber;
         this.emptyArray();
         this.shouldRun = true;
@@ -351,12 +306,6 @@ export class ProviderComponent {
 
     onOnlyActive() {
         this.onlyActive = !this.onlyActive;
-        console.log("change active filter, onlyActive=", this.onlyActive);
-        if (this.onlyActive == true) {
-            console.log("listing only active providers");
-        } else {
-            console.log("listing all providers");
-        }
         this.getProvidersByPageNumAndState(this.pageNumber);
     }
 
@@ -386,34 +335,29 @@ export class ProviderComponent {
 
     showUploading() {
         this.upload = !this.upload;
-        console.log("uploading", this.upload);
     }
 
     transform(bytes) {
         if (bytes == 0) return '0 Bytes';
-        var k = 1000;
-        var sizes = ['Bytes', 'KB', 'MB', 'GB'];
-        var i = Math.floor(Math.log(bytes) / Math.log(k));
+        let k = 1000;
+        let sizes = ['Bytes', 'KB', 'MB', 'GB'];
+        let i = Math.floor(Math.log(bytes) / Math.log(k));
         return (bytes / Math.pow(k, i)).toFixed(1) + ' ' + sizes[i];
     }
 
     setAttachments(event) {
-        console.log("set attachments ", event);
         this.selected.attachments = event;
     }
 
     setNewProviderAttachments(event) {
-        console.log("set attachments ", event);
         this.newProvider.attachments = event;
     }
 
     setLogo(event) {
-        console.log("Setting logo to provider " + this.selected.providerId, event);
         this.selected.logoUrl = event;
     }
 
     setNewProviderLogo(event) {
-        console.log("Setting logo to new provider", event);
         this.newProvider.logoUrl = event;
     }
 }
