+++ web/src/main/resources/public/src/app/user/provider/provider_type/service/provider-type.service.ts
@@ -14,40 +14,33 @@ import ApiService = require("../../../../../shared/services/api.service");
 export class ProviderTypeService {
     private url = ApiService.serverUrl + '/restful/providertype/';
 
-    constructor(private _http:Http){
+    constructor(private _http: Http) {
     }
 
-    getProviderTypes() : Observable<any>{
-        console.log("get all provider types inside service");
-        console.log("sending http GET");
-        return  this._http.get(this.url)
+    getProviderTypes(): Observable<any> {
+        return this._http.get(this.url)
             .map(res => res.json())
-            .catch((err)=>Observable.throw(err));
+            .catch((err) => Observable.throw(err));
     }
 
-    getProviderTypeById(id: number) :  Observable<any> {
+    getProviderTypeById(id: number): Observable<any> {
         return this._http.get(this.url + id)
             .map(res => res.json())
-            .catch((err)=>Observable.throw(err))
+            .catch((err) => Observable.throw(err))
     }
 
 
-    deleteProviderTypeById(providerTypeId:number) {
+    deleteProviderTypeById(providerTypeId: number) {
         let url = this.url + providerTypeId;
-        console.log('delete provider by id: ' + providerTypeId);
-        console.log("sending http DELETE to " +url);
         return this._http.delete(url)
             .map(res => res.json())
-            .catch((err)=>Observable.throw(err))
+            .catch((err) => Observable.throw(err))
     }
 
-    editAndSaveType(type:ProviderType) {
-            console.log('updating provider type with id: ' + type);
-            console.log("sending http PUT to " +this.url + type);
-            console.log("json obj: " + JSON.stringify(type));
-            return this._http.put(this.url + type, JSON.stringify(type),)
-                .map(res => res.json())
-                .catch((err)=>Observable.throw(err))
+    editAndSaveType(type: ProviderType) {
+        return this._http.put(this.url + type, JSON.stringify(type),)
+            .map(res => res.json())
+            .catch((err) => Observable.throw(err))
     }
 
 }
